Added stripReferences option
